+++ src/store/store.hook.js
@@ -1,7 +1,6 @@
 var logger         = require('../logger.js');
 var transaction    = require('./store.transaction.js');
 var lunarisExports = require('../exports.js');
-var debug          = require('../debug.js');
 
 transaction.registerHookFn(pushToHandlers);
 
@@ -174,8 +173,6 @@ function pushToHandlers (store, hook, payload, transactionId, callback) {
   if (!_storeHooks) {
     return callback();
   }
-
-  debug.log(store.name, debug.NAMESPACES.HOOKS, 'emits -> ' + hook);
 
   if (transactionId && (hook === 'error' || hook === 'errorHttp')) {
     transaction.addErrorEvent(transactionId);